+++ app/setup_handlers.go
@@ -6,6 +6,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/module"
 	"github.com/cosmos/cosmos-sdk/x/upgrade/types"
 	crosschaintypes "github.com/zeta-chain/zetacore/x/crosschain/types"
+	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 )
 
 const releaseVersion = "v12.0.0"
@@ -17,7 +18,8 @@ func SetupHandlers(app *App) {
 		for m, mb := range app.mm.Modules {
 			vm[m] = mb.ConsensusVersion()
 		}
-		vm[crosschaintypes.ModuleName] = vm[crosschaintypes.ModuleName] - 1
+		vm[crosschaintypes.ModuleName] = 3
+		vm[observertypes.ModuleName] = 4
 		return app.mm.RunMigrations(ctx, app.configurator, vm)
 	})